requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
django-extensions==3.2.3
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-anymail==10.2
Pillow==10.1.0
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
pytest-django==4.7.0
factory-boy==3.3.0
django-debug-toolbar==4.2.0
requests==2.31.0
